+++ gameplay/elements/player_input_controller/player_input_controller.gd
@@ -5,6 +5,7 @@ class_name PlayerInputController
 @export var state_machine: StateMachine
 
 var idle_state
+var no_input_state
 
 var selected_unit: Unit
 
@@ -18,8 +19,13 @@ func _ready() -> void:
 	EventBus.player_turn_started.connect(func ():
 		state_machine.set_default_state_and_push(idle_state)
 	)
+	EventBus.player_turn_ended.connect(func ():
+		state_machine.set_default_state_and_push(no_input_state)
+	)
 	idle_state = IdleState.new(state_machine)
 	state_machine.set_default_state_and_push(idle_state)
+	no_input_state = NoInputState.new(state_machine)
+	state_machine.set_default_state_and_push(no_input_state)
 
 
 func select_unit_by_cell(cell: Vector2i):