+++ gameplay/enemy_input_controller.gd
@@ -3,5 +3,8 @@ extends Node
 func _ready() -> void:
 	EventBus.enemy_turn_started.connect(func():
 		print("BUahahah >:(")
-		GameplayController.end_enemy_turn()
+		var timer = get_tree().create_timer(3.0)
+		timer.timeout.connect(func():
+			GameplayController.end_enemy_turn()
+		)
 	)